requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
bidict==0.23.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
django==5.1.5
docutils==0.21.2
fastapi==0.115.7
frozenlist==1.5.0
h11==0.14.0
htmx==0.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.5
markdown2==2.5.3
markupsafe==3.0.2
multidict==6.1.0
nicegui==2.10.1
orjson==3.10.15
propcache==0.2.1
pscript==0.7.7
pydantic==1.10.21
pygments==2.19.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pyyaml==6.0.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sseclient==0.0.27
sseclient-py==1.8.0
starlette==0.45.3
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.4
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
